+++ docker-compose.yml
@@ -9,6 +9,17 @@ services:
     environment:
       MONGO_INITDB_ROOT_USERNAME: root
       MONGO_INITDB_ROOT_PASSWORD: root_password
+  mongo-express:
+    image: mongo-express
+    restart: always
+    ports:
+      - 8081:8081
+    environment:
+      ME_CONFIG_MONGODB_ADMINUSERNAME: root
+      ME_CONFIG_MONGODB_ADMINPASSWORD: root_password
+      ME_CONFIG_MONGODB_URL: mongodb://root:root_password@mongo:27017/
+    depends_on:
+      - mongo
   backend:
     build: .
     ports:
