+++ src/util/tests/bcrypt_test.rs
@@ -4,5 +4,5 @@ use util::bcrypt::{password_hash, verify_password};
 fn bcrypt_test() {
     let password_hash = password_hash(&"123".to_string()).unwrap();
     let verify = verify_password(password_hash, &"123".to_string()).unwrap();
-    assert_eq!(true, verify);
+    assert_eq!(true.clone(), verify);
 }
